requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
llama-index>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
openai>=1.3.0
openai-whisper>=20231117
streamlit>=1.28.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pydantic>=2.5.0
